+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - [440fc19] [Feature] Add gemini-2.0-flash-thinking-exp
 - [e258b78] [Feature] Anthropic Prompt caching fully supported
 
+> Improvements:
+- [62d5556] [Improvement] Create mcp_config from sample if missing
+
 > Bug Fixes:
 - [4488ec1] [Bugfix] Trigger mcp servers shutdown before exit
 - [6d4d94e] [Bugfix] Skip asking to save chat on exit if disabled in config